requirements.txt
easyocr==1.6.2
Pillow==9.5.0
streamlit==1.21.0
transformers==4.28.1
sentencepiece==0.1.98